fix(schema): align Drift with Supabase — 3 critical compatibility fixes
1. organizations: add missing status + company_type columns (Drift + migration v17)
2. users.store_id: make nullable to support org-level admins (Drift + migration v18)
3. stockQty/minQty: migrate int→double across entire codebase for fractional
   quantity support (Drift, models, DTOs, DAOs, screens, tests)
   + Supabase migration to ALTER 9 columns to double precision

Also: regenerated all .freezed.dart and .g.dart files

+++ alhai_core/lib/src/dto/analytics/slow_moving_product_response.dart
@@ -38,7 +38,7 @@ class SlowMovingProductResponse {
       productName: productName,
       categoryName: categoryName,
       daysSinceLastSale: daysSinceLastSale,
-      stockQty: stockQty,
+      stockQty: stockQty.toDouble(),
       stockValue: stockValue,
       suggestedDiscount: suggestedDiscount ?? 0,
       lastSaleDate: lastSaleDate != null ? DateTime.tryParse(lastSaleDate!) : null,
