+++ packages/alhai_shared_ui/lib/src/screens/products/products_screen.dart
@@ -1184,7 +1184,7 @@ class _ProductListCardState extends State<_ProductListCard> {
                           ),
                           const Spacer(),
                           Text(
-                            AppLocalizations.of(context)!.stockCount(widget.product.stockQty),
+                            AppLocalizations.of(context)!.stockCount(widget.product.stockQty.toInt()),
                             style: Theme.of(context).textTheme.bodySmall?.copyWith(
                               color: colorScheme.onSurfaceVariant,
                             ),
